requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
APScheduler==3.11.0
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
discord==2.3.2
discord.py==2.5.2
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.2.0
propcache==0.3.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.3
tzlocal==5.3.1
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.18.3
